Replace remote_config directory with platform
The platform directory has an ${os}_${arch} subdirectory
per target platform. Each subdirectory contains a platform target
with the same name. For example, we can use
--platforms=//platform/linux_amd64. The remote C++ toolchain
configuration goes in here.

These platforms are not used yet. They were mentioned in .bazelrc
(removed) but not in a way that worked.

For PE-1127