Use hermetic Python. (#228)
Without hermetic Python, it's possible for the requirement lock to be generated using a different interpreter from the one that's actually going to run the test (e.g. in remote execution). pytest has a dependency where it mattered that the GitHub runner was Python 3.11 and the EngFlow executor was Python <3.11. And this is good practice anyway.

Resolves PE-1188