Binary input java example (#363)
## Binary Input Example
The goal of this example project is to test the performance of Engflow's
remote execution and caching service based on the number of input binary
files in the dependency graph. The project contains a `genrule` that
generates a specified number of Java binaries for the `genbinary` Java
library, which are then listed as dependencies in the main binary. The
`Main.java` file loops through each generated class and calls its
`greetNum` method.

### How It Works

1. **Generation of Java Binaries:**
- The `genrule` in the `BUILD` file generates a specified number of Java
classes (`Hello1.java`, `Hello2.java`, ..., `HelloN.java`).
- Each generated class contains a `greetNum` method that prints a unique
message.

2. **Main Class Execution:**
- The `Main.java` file in `binaryinput` dynamically loads each generated
class using reflection.
- It then invokes the `greetNum` method of each class, printing the
corresponding message.

---------

Co-authored-by: Andrés Felipe Barco Santa <andres@engflow.com>
Co-authored-by: Andrés Felipe Barco <andrez.fbs@gmail.com>